moved Color class to DataHolderLib
