+++ Applications/DataExplorer/DataView/StratView/StratView.cpp
@@ -22,7 +22,7 @@ StratView::~StratView()
 }
 
 void StratView::setStation(GeoLib::StationBorehole* station,
-                           std::map<std::string, GeoLib::Color>* stratColors)
+                           std::map<std::string, DataHolderLib::Color>* stratColors)
 {
     _scene = new StratScene(station, stratColors);
     setScene(_scene);
