+++ Applications/DataExplorer/VtkVis/VtkPointsSource.cpp
@@ -12,10 +12,6 @@
  *
  */
 
-// ** INCLUDES **
-// GeoLib
-#include "Color.h"
-
 #include "VtkPointsSource.h"
 
 #include <logog/include/logog.hpp>
@@ -32,6 +28,8 @@
 #include <vtkCellData.h>
 #include <vtkProperty.h>
 
+#include "Applications/DataHolderLib/Color.h"
+
 vtkStandardNewMacro(VtkPointsSource);
 
 VtkPointsSource::VtkPointsSource()
@@ -40,7 +38,7 @@ VtkPointsSource::VtkPointsSource()
     _removable = false; // From VtkAlgorithmProperties
     this->SetNumberOfInputPorts(0);
 
-    const GeoLib::Color c = GeoLib::getRandomColor();
+    const DataHolderLib::Color c = DataHolderLib::getRandomColor();
     GetProperties()->SetColor(c[0] / 255.0,c[1] / 255.0,c[2] / 255.0);
 }
