+++ Applications/DataExplorer/VtkVis/VtkStationSource.h
@@ -15,13 +15,11 @@
 #ifndef VTKSTATIONSOURCE_H
 #define VTKSTATIONSOURCE_H
 
-// ** INCLUDES **
-#include "VtkAlgorithmProperties.h"
 #include <vtkPolyDataAlgorithm.h>
 
-// GeoLib
-#include "Color.h"
 #include "Station.h"
+#include "Applications/DataHolderLib/Color.h"
+#include "VtkAlgorithmProperties.h"
 
 /**
  * \brief VTK source object for the visualisation of station data (including boreholes)
@@ -36,7 +34,7 @@ public:
 
     /// Returns the colour lookup table generated for boreholes.
     /// This method should only be called after the colour lookup table has actually been build (via RequestData() or setColorLookupTable()).
-    const std::map<std::string, GeoLib::Color>& getColorLookupTable() const
+    const std::map<std::string, DataHolderLib::Color>& getColorLookupTable() const
         { return _colorLookupTable; }
 
     /// Returns the type of observation site represented in this source object
@@ -44,8 +42,8 @@ public:
         { return static_cast<GeoLib::Station*>((*_stations)[0])->type(); };
 
     /// Sets a predefined color lookup table for the colouring of borehole stratigraphies
-    int setColorLookupTable(const std::string &filename)
-        { return GeoLib::readColorLookupTable(_colorLookupTable, filename); }
+//    int setColorLookupTable(const std::string &filename)
+//        { return GeoLib::readColorLookupTable(_colorLookupTable, filename); }
 
     /// Sets the stations as a vector
     void setStations(const std::vector<GeoLib::Point*>* stations) { _stations = stations; }
@@ -72,7 +70,7 @@ protected:
 
     /// The colour table for stratigraphic data. This table is either set using the setColorLookupTable() method or is generated
     /// automatically with random colours while creating the VtkStationSource-object.
-    std::map<std::string, GeoLib::Color> _colorLookupTable;
+    std::map<std::string, DataHolderLib::Color> _colorLookupTable;
 
 private:
     std::size_t GetIndexByName( std::string const& name );
