+++ Applications/DataHolderLib/Color.h
@@ -20,22 +20,22 @@
 #include <map>
 #include <string>
 
-namespace GeoLib
+namespace DataHolderLib
 {
-using Color = std::array<unsigned char, 3>;
+using Color = std::array<unsigned char, 4>;
 
 /// Returns a random RGB colour.
 Color getRandomColor();
 
 /// Reads a color-lookup-table from a file and writes it to a map.
-int readColorLookupTable(std::map<std::string, GeoLib::Color> &colors, const std::string &filename);
+//int readColorLookupTable(std::map<std::string, GeoLib::Color> &colors, const std::string &filename);
 
 /// Uses a color-lookup-table (in form of a map) to return a colour for a specified name. If the name is not
 /// in the colortable a new entry is created with the new name and a random colour.
-Color const getColor(const std::string &id, std::map<std::string, GeoLib::Color> &colors);
+Color const getColor(const std::string &id, std::map<std::string, DataHolderLib::Color> &colors);
 
 /// Convenience function to use the getColor method with numbers as identifiers.
-Color const getColor(double id, std::map<std::string, GeoLib::Color> &colors);
+Color const getColor(double id, std::map<std::string, DataHolderLib::Color> &colors);
 } // namespace
 
 #endif /* COLOR_H_ */
